Moved obsolete classes to obsoletes-v8
