+++ src/NServiceBus.Core/Persistence/PersistenceConfig.cs
@@ -11,7 +11,7 @@
         /// <summary>
         /// Configures the given persistence to be used.
         /// </summary>
-        /// <typeparam name="T">The persistence definition eg <see cref="InMemoryPersistence" />, NHibernate etc.</typeparam>
+        /// <typeparam name="T">The persistence definition eg <see cref="LearningPersistence" />, NHibernate etc.</typeparam>
         /// <param name="config">The <see cref="EndpointConfiguration" /> instance to apply the settings to.</param>
         public static PersistenceExtensions<T> UsePersistence<T>(this EndpointConfiguration config) where T : PersistenceDefinition
         {
@@ -23,7 +23,7 @@
         /// <summary>
         /// Configures the given persistence to be used for a specific storage type.
         /// </summary>
-        /// <typeparam name="T">The persistence definition eg <see cref="InMemoryPersistence" />, NHibernate etc.</typeparam>
+        /// <typeparam name="T">The persistence definition eg <see cref="LearningPersistence" />, NHibernate etc.</typeparam>
         /// <typeparam name="S">The <see cref="StorageType" />storage type.</typeparam>
         /// <param name="config">The <see cref="EndpointConfiguration" /> instance to apply the settings to.</param>
         public static PersistenceExtensions<T, S> UsePersistence<T, S>(this EndpointConfiguration config) where T : PersistenceDefinition
@@ -38,7 +38,7 @@
         /// Configures the given persistence to be used.
         /// </summary>
         /// <param name="config">The <see cref="EndpointConfiguration" /> instance to apply the settings to.</param>
-        /// <param name="definitionType">The persistence definition eg <see cref="InMemoryPersistence" />, NHibernate etc.</param>
+        /// <param name="definitionType">The persistence definition eg <see cref="LearningPersistence" />, NHibernate etc.</param>
         public static PersistenceExtensions UsePersistence(this EndpointConfiguration config, Type definitionType)
         {
             Guard.AgainstNull(nameof(config), config);
