+++ src/NServiceBus.Core/Persistence/PersistenceExtensions.cs
@@ -10,7 +10,7 @@
     /// This class provides implementers of persisters with an extension mechanism for custom settings for specific storage
     /// type via extension methods.
     /// </summary>
-    /// <typeparam name="T">The persister definition eg <see cref="InMemoryPersistence" />, etc.</typeparam>
+    /// <typeparam name="T">The persister definition eg <see cref="LearningPersistence" />, etc.</typeparam>
     /// <typeparam name="S">The <see cref="StorageType" />storage type.</typeparam>
     public class PersistenceExtensions<T, S> : PersistenceExtensions<T>
         where T : PersistenceDefinition
@@ -28,7 +28,7 @@
     /// This class provides implementers of persisters with an extension mechanism for custom settings via extension
     /// methods.
     /// </summary>
-    /// <typeparam name="T">The persister definition eg <see cref="InMemoryPersistence" />, etc.</typeparam>
+    /// <typeparam name="T">The persister definition eg <see cref="LearningPersistence" />, etc.</typeparam>
     public class PersistenceExtensions<T> : PersistenceExtensions where T : PersistenceDefinition
     {
         /// <summary>
